chore: fix eslint issues
