+++ src/design-system/index.js
@@ -1,2 +1,1 @@
-export * from './theme';
 export * from './components';
