+++ src/design-system/theme/hooks/index.js
@@ -1,0 +1,2 @@
+export { useThemeSwitcher } from './useThemeSwitcher';
+export { useAppTheme } from './useAppTheme';
