--- src/design-system/theme/index.js
@@ -1,1 +1,0 @@
-export { useAppTheme, useThemeSwitcher } from './ThemeProvider';
